Add additional asserts and move TrackVirtual test
